refactor(libsinsp/test): split libsinsp UTs into multiple files
